refactor(root): Reworked file structure [skip ci]
