+++ api/configs/index.js
@@ -1,0 +1,9 @@
+import config from './configuration';
+import database from './database';
+import middlewares from './middlewares';
+
+export {
+  config,
+  database,
+  middlewares,
+};
