+++ api/utils/index.js
@@ -1,11 +1,3 @@
-import config from './configuration'
-import database from './database';
 import { filterBody } from './filterBody';
-import middlewares from './middlewares';
 
-export {
-  config,
-  database,
-  filterBody,
-  middlewares,
-}
+export { filterBody }
